walk strings in the dom tree
Fixes a test failure in ADWolf (`form-object-create-destroy`). The
error happened with a `multi-email-box` with several
emails. `adwolf2006::while-collapsed` rendered the `multi-email-box`,
which returned a list like: `(<a ...> ", " <a ...>)`

`adwolf2006::while-collapsed` then used the dom walker to call
`adwolf2006::clean-names-and-ids` on every element of the freshly
rendered `multi-email-box`, which has a literal string in it.

Not sure if this is a bug somewhere else.